warning消し

+++ src/Solver/ikerukanaDPSolver.cpp
@@ -48,12 +48,12 @@ namespace solver {
             que.pop();
             assert(Graph[que_front.node_num].size() <= 4);
             for (const size_t& dist_node: Graph[que_front.node_num]) {
-                if (direction_table[que_front.node_num][dist_node] == que_front.direction) continue;
+                if ((size_t)direction_table[que_front.node_num][dist_node] == que_front.direction) continue;
 
-                bool& dp_target_elm = dp[que_front.remaining_move-1][dist_node][direction_table[dist_node][que_front.node_num]];
+                bool& dp_target_elm = dp[que_front.remaining_move-1][dist_node][(size_t)direction_table[dist_node][que_front.node_num]];
                 if ((int)que_front.remaining_move-1 > 0) {
                     if (not dp_target_elm) {
-                        que.push(BFS_Status{que_front.remaining_move-1, dist_node, direction_table[dist_node][que_front.node_num]});
+                        que.push(BFS_Status{que_front.remaining_move-1, dist_node, (size_t)direction_table[dist_node][que_front.node_num]});
                     }
                 } else {
                     assert((int)que_front.remaining_move-1 >= 0);
